implementa interface de métodos abstratos nas views de ativos
