+++ src/views/MenuNft.java
@@ -76,7 +76,7 @@ public class MenuNft {
                 break;
 
             case 7:
-                MenuUsuario.salvarArquivo(idUsuario);
+                MenuAtivoInterface.salvarArquivo(idUsuario);
                 break;
             case 8:
