+++ src/views/MenuUsuario.java
@@ -4,6 +4,8 @@ import controllers.AtivosController;
 import controllers.GerenciarUsuario;
 import helpers.Console;
 
+import static views.MenuAtivoInterface.finalizar;
+
 public class MenuUsuario {
 
     private static void exibirMenuUsuario() {
@@ -102,22 +104,7 @@ public class MenuUsuario {
         }
     }
 
-    private static void finalizar(String idUsuario) {
-        String salvar = Console.lerString("Deseja salvar as alterações antes de finalizar (S/N)? ");
-        if (salvar.equals("S") || salvar.equals("s")) {
-            salvarArquivo(idUsuario);
-        }
-        System.out.println("\nSistema finalizado...");
-    }
 
-    public static void salvarArquivo(String idUsuario) {
-        try {
-            AtivosController.salvarArquivo(idUsuario);
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-            return;
-        }
-    }
     
     // private static void login() {
